17-08-2020. Ajout de quelques modifs

+++ src/servlet/DeleteOrders.java
@@ -2,6 +2,7 @@ package servlet;
 
 import entities.AdsEntity;
 import entities.OrdersEntity;
+import entities.UsersEntity;
 import enumeration.EnumOrderStatut;
 import forms.AdsForm;
 import forms.BasketCrudForm;
@@ -13,6 +14,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 
@@ -21,23 +23,36 @@ import java.io.IOException;
 public class DeleteOrders extends HttpServlet {
 
     public static final String VUE = "/WEB-INF/deleteOrders.jsp";
+    public static final String URL_REDIRECT = "listOrders";
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-        this.getServletContext().getRequestDispatcher(VUE).forward(request, response);
+        response.sendRedirect(URL_REDIRECT);
+//        this.getServletContext().getRequestDispatcher(VUE).forward(request, response);
 
     }
 
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        OrdersForm ordersForm = new OrdersForm();
-        ordersForm.deleteOrder(request);
+        // récupération de la session
+        HttpSession session = request.getSession();
 
-        request.setAttribute("ordersForm", ordersForm);
-//        response.sendRedirect(URL_REDIRECTION);                    // Cas d une redirection
+        UsersEntity usersEntity = (UsersEntity) session.getAttribute("UserEntity");
 
-        this.getServletContext().getRequestDispatcher(VUE).forward(request, response);
+        if (usersEntity != null) {
+            if (!usersEntity.getRolesByIdRoles().getLabel().equals("Admin")) {
+                response.sendRedirect(URL_REDIRECT);
+            } else {
+                OrdersForm ordersForm = new OrdersForm();
+                ordersForm.deleteOrder(request);
 
+                request.setAttribute("ordersForm", ordersForm);
+
+                this.getServletContext().getRequestDispatcher(VUE).forward(request, response);
+
+            }
+        } else {
+            response.sendRedirect(URL_REDIRECT);
+        }
     }
 }
